fix task and errors

+++ app/Http/Controllers/V1/Task/TaskController.php
@@ -27,7 +27,7 @@ class TaskController extends Controller
         $user = Auth::user();
 
         if (! $user) {
-            return response()->json(['message' => __('messages.user.Inaccessibility')]);
+            return response()->json(['message' => __('messages.user.Inaccessibility')] , 401);
         }
 
         return IndexTaskResource::collection($this->taskrepo->index());
@@ -44,7 +44,7 @@ class TaskController extends Controller
         $group = null;
 
         if (! $auth) {
-            return response()->json(['message' => __('messages.user.Inaccessibility')]);
+            return response()->json(['message' => __('messages.user.Inaccessibility')] , 401);
         }
         
         if($request->group_id !== null){
@@ -65,7 +65,7 @@ class TaskController extends Controller
         $user = Auth::user();
 
         if (! $user) {
-            return response()->json(['message' => __('messages.user.Inaccessibility')]);
+            return response()->json(['message' => __('messages.user.Inaccessibility')] , 401);
         }
 
         return new ShowTaskResource($task);
@@ -73,12 +73,17 @@ class TaskController extends Controller
 
     public function update(Task $task, UpdateTaskRequest $request)
     {
-        $auth = Auth::id();
+        $auth = Auth::user();
         $group = $task->group_id;
 
-        if (! $auth) {
-            return response()->json(['message' => __('messages.user.Inaccessibility')]);
+        if (! $auth->id) {
+            return response()->json(['message' => __('messages.user.Inaccessibility')] , 401);
         }
+
+        if($auth->id !== $task->owner_id){
+            return response()->json(['message' => 'عدم دسترسی'] , 403);
+        }
+        
         
         if($group){
             $group = Group::find($task->group_id);
@@ -104,6 +109,16 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        $auth = Auth::user();
+
+        if (! $auth->id) {
+            return response()->json(['message' => __('messages.user.Inaccessibility')] , 401);
+        }
+
+        if($auth->id !== $task->owner_id){
+            return response()->json(['message' => 'عدم دسترسی'] , 403);
+        }
+
         $error = $this->taskrepo->delete($task);
         if ($error === null) {
             return response()->json(['message' => __('messages.task.delete.success')], 200);
